+++ CHANGELOG.md
@@ -4,8 +4,11 @@
 -->
 
 # Changelog
+## __WORK IN PROGRESS__
+* Config parameters are rendered in a preview
+
 ## 0.0.6 (2023-01-23)
-* Feat: Quick fix for properties that unnecessarily override `$import`ed properties
+* Added a quick fix for properties that unnecessarily override `$import`ed properties
 
 ## 0.0.5 (2023-01-23)
 * Fix: Decorations could be shown in wrong editor